commit water level

+++ src/main/java/com/neo/nbdapi/rest/vm/WaterLevelVM.java
@@ -17,7 +17,7 @@ public class WaterLevelVM implements Serializable {
 
     private Long tsId;
 
-    private Integer value;
+    private Float value;
 
     private Integer status;
